pyproject.toml
[project]
name = "playground-copilot-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "github-copilot-sdk>=0.1.25",
    "ipython>=9.10.0",
]
